Added PixiJS and rendered map sprite
PixiJS is a 2D WebGL based rendering library
It fits weel for the game, as it is not a large
game library that would have lots of unnecessary
features. Since all the client does is render the
game updates and menus, PixiJS sounds perfect.
It renders everything to HTML canvas which
should have good performance.

In this commit I have added a "hello world" style
render of a tile that could be use for a map.