Added algoritm to generate path through map
This path is the path that the enemis will take.
The algorithm works by dividing each map square into a sub
grid of size given as a parameter. Each enemy will enter the
subgrid, move to the center, and then move towards the next map
square. By dividing the subgrid it gives a relative position
that the client can render as it chooses.
The speed of an enemy will be how many sub-grid squares it can
move through in one tick.