Added rendering optimisation
Now, on each server state update message, we hash the names of all the
neemies included and send this to the client. The client keeps track of
the previous hash. If there is no differene it means the same enemies
are present as the previous update, so no need to check if anything has
reached the end/been killed or new enemy spawned.

Watching performance shows that occasionally the rendering could be a
bit choppy when a high number of enemies were present. Now it seems
smooth at all times.