Added dragable towers
Added a tower sprite. One is placed in the "menu" and when clicked and
dragged it creates a sprite that can be placed on the map. Once placed
is fixed and a message is sent to the server which updates its map with
the new tower. Whilst the sprite is being moved, a message is send at
each position which will eventually be broadcast to the other clients so
they can see the sprite being moved. It will be a temporary sprite that
will be removed once the map is updated on the server.