Tidy up part 2
Decided to keep all rendering work in one file. Was thinking of
splitting out nto separate files, but there is so much shared state on
the rendered that though best to keep it as is. As such organised into
different types of function.

Also passed through set up info such as grid propertied from the server.