Made towers follow enemies as they move
This has been implemented by storing the range of squares that t tower
can shoot to based on its range. THese are pre-computed on the towers
placement. The first enemy to be detected in one of these squares is the
one the tower will shoot at. The angle for the tower sprite to face is
calculated using the main grid coordinates of the tower and the enemy.

This is all calculated in the server, and a message is sent to the
client and contains the angle of each tower.