Added collision detection between bullets and enemy objects
Each bullet checks its position to the enemy hitbox (a circle of given
radius). If the bullet is within that radius it is a hit. If not, start
moving back down the path of the bullet and using points along that
line. It checks from, teh current position to the position in the
previous tick. Currently set to make 4 checks, including the ceck for
initial position.

If a hit is detected, the bullet is removed.

Also includes some changes intended for the previous commit.