Sent enemy/bullet collision update to renderer.
If the client sees that an enemy has been hit, it tints the sprite for a
short period of time. Ideally, would lighten the sprite for a bit.
Having a bit of trouble appling filters to the black parts of the sprite
- presumably because thery are 0x000000.

This will do for now - just want a visual representation of a hit.