Added range circle when dragging tower on client
This is a circle that indicates how far a bullet from this tower can travel.
The circle remains in the container once the tower has been dropped.

Also added the config for each tower as a json file that can be shared
by the server and client. jQuery has been included to help retrieve the
json.