New game now starts separate game for each user
Creates a room whti its own map, enemies and towers. Join game will
later allow other players to enter that game. Updates are sent to any
client in that room.

Game state and functions have been moved to a new class which holds the
game state for a single room.