Added tower menu
A menu that allows a user to drag and drop a tower onto the map has been
added. It is currently just places to the right of the map and has three
of the same tower.

I have also started using the PIXI container bound (which is the square
around all the sprites in the container) as as way to do mouseover
detection. This is used to determine whether to snap the tower to grid
or not.