Made mobile browser keyboard appear
Required adding an input HTML element. Putting focus on that element
makes the keyboard appear on mobile. Bit hacky but it works, and seems
to be the method used online.

Currently only works on Android Firefox, does not work on Chrome.

Also added a debug message type, that prints anything to the server
console. This is useful when no mobile dev tools console.