Added styling to canvas presentation
The game window has been centered on the screen. A border has also been
added.

A bug was also fixed with the http server. When the file contents is
read to pe put in the response, a promise is created, with the file
being send once it is fulfilled. There was a bug where the MIME type
given in the response was whatever value the MIME_type variable had at
the time of promise fulfilment - instead of the value at promise
creation. To fix this, a function is called which creates its own scope.
The mime type variable is set within the function, the file is read and
the promise is created. Then when the file is ready to be sent, the mime
type from the local scope is used - this is the correct mime type.

This issue went unnoticed until now becae the majority of mime types
used were text/javascript. But with the addition of text/css I started
seeing issues where the css was deliverd with the wrong mime type. Since
it occasionally worked, I eventually realised it was a race condition on
the server.