Restructured the tower menu icons logic
Now, once the icon is dragged a temporary icon is created which can be
moved and placed. Once placed, the actual tower is added and the
temporary sprite is removed.

The code has been decoupled into smaller functions with more general
usage.

The tower info is also displayed when the menu icons are clicked.

Note that ther is currently an issue with placing a tower when the
cursor is not over the tower (such as if it is on a path). This will be
fixed in a later commit.