Improved the tower aiming algorithm
Towers now look ahead to the possible positions of the enemy that they
are aiming at. For each one, the position of the bullet in that number
of ticks, whilst moving toward that position, is calculated. If there
would be a collision in that many ticks, then we aim for that position,
knowing that the tower will move into the bullet (assuming no other
tower gets in front, which is perfectly possible, legal, and expected).