Changed the way enemies are iterated through in moveEnemies
Some enemies were moving extra times due to the way we iterate through
the enemies in each grid square. Since it was iterating forward,
potentially moving an enemy, then moving all enemies in the next grid,
the enemies that cross the bondary mve twice.

Also want to stop reordering enemies as they are being moved for the
current square, because again would move twice if an enemy overtook the
next one.