Moved logic from bullet to tower
Didn't make sense to set a target on the bullet and workout speed and
trajectory from that. Makes more sense to have it just a simple object
that the tower tells where to go.

Also have added a simple "shoot" interface on the tower that will call a
certain bullet pattern depending on the tower type. THis means the game
does not need to know that behaviour of the tower when it shoots, which
is neater.

Also simplified the trigonometry of the bullet speed using the angle and
parametric coordinates.