Added custom forEach style loops for iterating through all enemies
This cleans up code a bit, and make it easy to iterate over all the
enmiesindepndendt of where they are.

Also included way to iterate in reverse, since moving enemy position,
or deleting enemy would affect the next enemy come to.

Also ensures that any enemy added to the map data structure is in step
order within it square. This is easier that reordering on move, as
intevitable some future change will move an enemy and would need to
manually reorder again then.