Send only limited map information to client to prevent slowdown on tower placement
There was a pause when a tower is placed. THis was cause by 100MB or so
of data being sent to the client when the tower is placed, as part of
the map stucture update call. To fix this a mimial representation of the
map is sent - contains value only. The only other information that was
used was the colour of a tower. This has been set to the the users
tower, but will be reverted back when other user information is
stored in the client soon.

Also fixed an issue where the tower update message was not checking the
has of all tower names, so looped through each time unecessarily.