Moved tower sprites ligic into a separate class, and added a toolbar variant for the tower icons
This should make the renderer code much cleaner. It removes a lot of code
from the ever growing renderer file. It also changes the mchanic of the
tower button. now, there is just a sprite that when hovered over will
produce a tower sprite which can be dragged and placed as normal. If the
user chooses to not move the tower, it stays on top of the towe sprite,
and another one is not produced on hover over.

Also had to move the dimensions of the view out to a file for global
access. This is nicer in general to be honest, and gives an easily
accessible place to change these if required. HOWEVER - it is currently
just harcoded to be the same as the server settings. A shared config is
needed.