Added linked component feature
This allows components to acess each other if they need to. For
instance, the tower that is produced by the tower menu component, when
hovered over, needs to show some information on the tower info toolbar.
Through the link, it can call a function "onTowerMenuPointerover". This
interface means the tower menu has no say in what happens when hovered
over, the tower info toolbar decides - this is goom the tower menu does
not really care or need to know. It just needs to know that this must
happen.

The tower menu component produces the tower icon, shoves on all the
events for itself and any linked components, and then drops the sprite
in to the sprite layer.