Adding aim behaviour buttons.
Aim behaviour dictates which enemy a tower will shoot in its range. Not
yet sure whether a player will be able to change this during the round.
Maybe just at the buying phase.

Just a few square graphics for now. Also only have two behaviours
implemented server side.

Server side, the tower now has a state object that records all the
properties that can be changed by a player.

When the button is clicked, it sends a message to ther server. The
server reads the type of the resource that is being updated and send the
data to the appropriate update call.