Rendered bullets on top of towers and enemies
Rendering on top of towers means bullets do not appear underneath other
towers. This was desired originally so the origin tower would not look
like it was shooting from the middle of the sprite, however looks
strange when a bullet travels under another tower. This solution works
by not sending a bullet to the clietnt to be rendered if it has not
travelled far enough away from the origin tower. When it has, it will be
sufficient enough to look like it is coming out of the end of the turret
gun.