Added more components to the lobby view
Included a view of the map that will be used for the game. The players
will be able to regenerate this map. Added a buttonm for that. ALso had
to ass a way to scale the map, since using the existing map code to
render it.

Moved the game code from the sprite handler.