Enemies can be triggered by pressing a button in game
This is the beginning of the rouns structure to the game. Enemies come
in waves, and rounds are started onca all players have spent their
money.

Also includes a bug fix for double kills. This meant that the counter
that is used to determine whether a round is over was off by one, so an
enemy was left lying around. I think the current implementation of just
comparing enemies "in" and "out" is a bit fragile and could be better.

Have decided to keep the game updates continual. This means that the
same update message is always sent, mainly for simplisity, but also so
that the current stats of the game can always be known. It will mean
that updates such as when another player is moving a tower can just be
part of that.