Improved the tower menu icons code
Original implementation was to produce a dragable tower sprite whenever
the mouse hovered over the icon. When it was done with it was removed,
and user would need to hover over the icon again to produce anothe
rdraggable tower. This was bad becuase it had inconsistent behaviour,
and did not work on mobile.

Have switched to use a system where a single draggable icon exists for
each tower type. It can be moved, but then a decision has been made
(confirm or bin, or dropped out of map) then it returns to its starting
location, ready to be picked up again.

Code and logic are both simpler.