Added info box UI component
THis is achieved using a 9 slice plane. This keeps the 4 corners fixed,
and only stretches the sides and middle. As such works very well for a
text/message box, as can have detail in the corners and can then resize
as required without stretching the details.

The imageis a simple graphic for the text box, can update with something
more detailed in teh future.

Also removed rendering tower info on the tower info container. As part
of the UI redesign, I would like to remove the fixed toolbars, and
instead the map view will be the only thing visible, and the UI
components will appear as needed.