Changed all components to inherit PIXI container
This is instead of having a main container and sub containers that are
added to. Also remove the dependnecy on the sprite handler in the object
construction. This is now only required for those classes which use the
actively clickable sprite. THis will be changed in the furture.

The tower menu was changed such that all sprites exist in the same root
container (the component). This means that the laye of the range
sprite is the same layer as the other tower menu components, instead of
above the toolbars etc as used to be. THis is ok for now as the toolbars
will be disappearing soon.