Replaced the updatables pattern to use a tick function pattern
Instead of passing the sprite handler into a component, can instead add
a "dynamic component" to the sprite handler. This dynamic conponent will
overrride the empty (virtual) tick function in the base class and it
will be called on every iteration of the render loop. This is a nicer
design pattern that makes conponents more unaware of how they are used.