Clean up transition code
Renamed to use "open" and "close" terminology. Also change the active
tower structure a bit so that it is removed and put into its own
container. Makes it easier to transition the non-active towers since
they are all shifted. Eventually will remove the active clickable tower
and keep it local.