Fixed tests broken by change in config
Config no longer ocntained the subgrid midpoint, so they failed becuase
they used that value. It was removed because it is a value derived from
another config option so is not necessary in the core config. Tests have
been kept the same, but a lcoal variable used instead. All tests now
pass.