Added the valley walls to the map
The sprite used depends on perspective for the vertical walls. If on the
right of the map, shows the rifht wall, but not left, and vice versa for
the left side.

The client sends information about each non-path tile. This information
tells the client which of its adjacent squares need to have a wall, and
the direction, using a number 0-7 to indicate the wall type, angle etc.