Changed the map sprite switch functions to allow any arguments
The switch function passed to makeSprite on a per spritetype basis is no
an anonymous function with one argument - thetransformation object. Any
other arguments are then directly inserted into the call to the type
specific switch. This means things like offset for the side walls does
not need to be passed into the switch function for the othe sprited -
same for sprite height.