Added new sprite for basic tower
This tower is the rock thrower/cannon. Just shoots a single bullet
(rock). All towers have the sprite and then the base. This base is a
grayscale part of the sprite, and is what gets tinted to show different
player colours. It is added as a child to the tower sprite so renders on
top of it.