Tidied up client tower sprite code
In preparation for adding each tower sprite and its aniations, moved the
tower sprite code into a separate set of classes. This allows each tower
type to eventuall have its own functions for animation, sprite layers,
etc, whilst still having some common behaviour.

Also updated to the json to use the tower type as the key and stopped
using the index based system which was fragile. Also included the
texture atlas name inthe tower json, to avoid changing it in the code.

The client to server application messaging had some redundant data, so
that was removed.

The new tower sprite class also means that there is a proper interface
for dealing with tower popup info and range sprites.