Added code for sprite shadows using masking, but removed all shadow code
This includes removeing shadows cast by the valley walls. I think it
actally looks better without it. Have left the code present in case I
change my mind. Something to focus on closer to the end of the project I
think. Better to keep the graphics simpler for now.

Same for the tower code, shadows looked ok but not good enough to make
it worthwile. The current implementation is not perfect, but is
acceptable.