Prepared for rotation of enemy sprites
THis involved changine the info the server sends to the client. Now sed
the type of tile and the order in which the enemy would travel i.e. ur
means enemy comes up then turns right. This is important because this is
a corner tile that si the same as the left down, but is ver different
when it comes to enemy movement.

In order to determine when to turn, will use this information as well as an
indicator that the sprite is near the centre of the subgrid. Will then
know to turh to the second direction in the tile type (or not turn if
straight tile).