Added some texture to the land
Placing randomly generated (pre-generated images) shapes, with slight
variations on colour, in random places gives a textured look. The colour
is so slight that it looks natural withouhg drawing detail.

The whole map is cached as a bit map to save processing. Quick profiling
in Task Manager shows up to 5% reduction in CPU usage which is good.