Added features to the map surface
The features are just cacti and rocks for now. They are rendered in
clumps which are randomly placed around the map. Four of them. Those
clumps have a high chance of rendering a sprite. A clump is a square on
the map and the few squares around it. It looks quite natural that way.

Featues can also not be placed on a square next to the map. This is to
avoid them spilling over into the path space.