Restructure of client tower code
Tower sprite forms (draggable tower menu, mapt tower etc) are now
represented by objects instead of a sprite with additional properties
attached on. This is a much more scalable and
clear form of representing the towers on the renderer.

Each tower type will have properties depenedent on the type parameter
given, and all use the same class for instantiating a tower of that
type. This is better than the previous implementation where a type fo
tower had its own class, with a class for each form of tower also.
Adding a new tower type may now require no code change if it uses the
basic method of rendering.

This also now mean that a link to the tower component object is not
required by the tower menu class for creating its towers. As such every
component is completely decoupled from any other components (or share a
parent in the case of the interactive game space componenents).

Also updated the tick function call to be powered by the server update
call insterad of a separate loop in the sprite renderer. This means a
dynamic componenet feature is not required any longer.