Prepared stage for game up boxes
Increased the space around the map, so that they do not lie directly on
the map itself.

Also added a possible ui box sprite.

The map object is not passed through to the interactive game area,
instead of passing through the 2d array from the client. The map
componenet holds that is is updated when server sends it through.