Updated the way the draggable tower moves
Previously if the mouse was outside the interactive space, the tower
would not move at all. Now, if the mouse is within the range from an x
or y perspective, that dimension of the towers position is still
updated. This gives a smoother feel when dragging the tower. Wihout this
movement, the tower would get stuck as soon as the mouse went outside the
boundary in either direction.