Changed deployed tower popup UI
This adds UI around the tower itself. Have added to "buffer" space
around the map and interactive area so that the menu options can appear
for every tower. However this uses up a lot of useful space and makes
the area look awkward and boxed off. I think having a UI menu where the
tower menu is would be better. Plus, the angle of the current menu icons
is too small and slanted, especially for the amount of info they need to
hold.