Switched to a single menu for control of deployed towers
This is preferrable than the current menu that hovers over each tower. I
prefer this method because it does not requre the extra space around the
towers, it allows more information to be displayed because it occupys
more space, and it is easier to append more options to if required in the
future.

There will be a single menu object, and it will be updated with the
options for the current tower, rather than storing a menu per tower.