Made selecting another deployed tower unclicks the currently selected one
Previously, multiple towers could be clicked at the same time. Now when
a deployed tower is clicked, if another one is already clicked then it
is closed and the menu bar is updated with info for the new tower.

Also made the clicked tower render below the other towers, because
otherwise they could not be clicked because the range sprite of the
clicked tower wa above them.