Updated the look of the range circle sprite
A dashed line fits the theme a bit more, and looks better. It also
rotates to make the whole view feel a bit less static.

Also render above other towers, since feels weird to have a single dash
hidden underneath. Plus, able to interact since there is no sprite in
the way anymore because it is just the dotted line.