Added options to the aim menu
This menu determines how the tower seleces the enemy to shoot. It sends
an update to the server which updates the state of the tower.

It is a switch menu, since only one aim behaviour can be seta at a time.
It defaults to first.

The server now also includes the aim behaviour in the tower state update
so that the client can display which one is currently selected.