Begun server restructure
This is the start of a set of work to reorganise and better structure
the server code. I feel this is needed because the code is messy and not
particularly well separated into modules of distinct functionality. It
is also not clear exactly what the behavious are when a client sends an
update.

This commit starts off by adding a session class, wich
represents a single game and the players involved. His is beneficial
because it creates a logical closed collection of objects relevant to
that one playing session. It also means that every event is not added to
the socket until it is legitimately added to a game.

By directly accesing the game object within the session, the behaviour
of the server remains unchanged from the clients point of view and the
game plays as normal.