Split server map class into map and map generator
The map generator will just be for creating a map. Once the game has
started that map will be passed to the game class which handles moving
all the elements within that map and progressing the game.

It's a logical separation since would never need to generate a map once
the game has begun.