Changed the way event names are used in client and server
Originally an enum was used, but due to inconsistencies in importing
files in node and client js, that enum could not be shared between the
two. The initial solution was to duplicate that enum across both client
and server files, but that was unnecessary duplication.

Instead, the enum has been removed and the string event names are used.
Part of the reason for the enum was to enforce that the even existed
which was quite nice, but I think a descriptive and well structured
set of events is a better way to manage it.