Inisialised WebSocket connection upon app load and pass to components
This is better than having to include a file which has a global socket
IO socket object. It is then pass to each view to mantain the connection
for the user.

Also is perferrable to have direct access to the socket object rather
then a thin wrapper in the form of the networking.js file.

These changes now makce it much clearer in the game and other views
exactly what messages are going back and forth between the client and
server. All messages that are sent to the server are sent from the new
event emitter object in the game view (for the game).