Changed the way the client gets map info
Getting the map now triggers a callback which returns the map data to the
client that called instead of using the set map event, which is
separate.

Fixed the map regeration not working due to not unding the bitmap cach
in the map renderer.

The set map event still exists, now in the lobby only. Once the game has
started the client will not need to have it's map reload triggered by the
server, it will only need to ask for it itself.