Changed how the game is updated, and the messages sent
Each component type, such as bullets, enemies can be updated separately
so more control can be given to exactly what is updated and sent.

The measure of whether a round is active hsa also been changed. A round
is active if there are still enemies left to kill and bullets on the
map. This avoids using two separate update functions.