Changes how the round state is maintained and reported to the client
In an effort to move from update based messaging to whole state
messaging i.e. where the client can render the whole game state with a
single update, the end round event has been removed. Instead, the client
will check if the round number sent by the server has changed.

Also the game itself now checks for whether the round is active or not,
and then updates the entities only if it's own internal state has active
set to true.