Fixed a couple of bugs with the start round button
One was setting all the children to interactive, which mean that the
text could be clicked but no events were assigned to it so the button
would do nothing. This has been changed so only the interactive prt of
the button has it's interactivity toggled during rounds.

Also changed the behaviour of lifting mouse when outside the button.
Originally this would trigger the button effect, but I think it is more
intuitive and useful to have the selection cancelled if the mouse is
released outside of the button sprite. This applies to menu buttons
in general, not just the start round button..