Changed how sprites are stored in a tower
There was a bug caused by setting the tint of the towers if the player
could not afford them. This was caused by the setting of the tint
property of the sprites that make up the tower. These sprites are stored
a bit weirdly - sometimes in a container and other times a single
sprite. This was introduced to handle spites like the shrapnel burst -
with multiple moving parts. The code has been changed so that all parts
fo the sprite - the main body and the coloured part are in their own
containers for consistency. Also, a method to set tint was added, as
this was a commonly used bit of functionality.

The bug has been fixed and now the coloured part retains its colour
after a tower is available to buy again.