Changed how player ID and colour config is set on the client
To ensure the colour of the players tower is set for all menu towers and
deployed towers, it is set in the config object when the player starts
the game. In future it will be configurable.

The config is also now shared between components, so the player ID no
lnger has to come from the global state import.