Changed how the tower basic shoot animation works
Originally it was moving the whole sprite, this meant the range circle
also moved but it should not. The sprite container is now the thing that
moves. This also means we do not need the store the starting position,
since the sprite container moves relative to it's parent which is the
towe robject.