Changed the order of rendering of some game components
Previously the tower menu towers appeared underneath the already placed
towers which meant the buttons to confirm placement could be below an
existing tower and therefor not clickable. This was because the towers
component was added to the game space after the tower menu, so is
rendered above it. This has ben rectified so that the order is correct.
Also changed is the arguments that the game space takes. To ensure that
the required components are present and can be ordered appropriately,
the towers, bullets and enemies componenets are required to create the
game space.