Display tower information when user hovers over tower in menu
To help the user get more information about a tower, render some
information about it. This includes a description, cost, and the type of
tower.

I think that it could look more appealing that it currently does.
Possibly more colours on the menu items, or a better font. Is very
simple at the moment.