Added a new enemy type
This new enemy is another beetle, smaller and faster than the current
one, and green in colour. It was created using the same sprites and
animation, but then was scaled down and the framerate increased on
export from Spriter. This is a fast way to create such a similar
sprite, although may become unyieldly in the future if the sprites need
to be edited in any way.