Added animation for when bullet hits enemy
This first required sending to the client the angles of collision
between a bullet and the enemy. The client then renders these a suitable
distance form the centre of the enemy sprite at the relevant angle. An
animation has been created in Spriter. It is a simple dust explosion
that loops once.

To facilitate this animation, an enemy class was created. It holds the
enemy sprite and has the collision animation added as a child.