Make map start and end on middle row
So that the town sprte can fit in between the counters on the right hand
side, need to ensure the map finishes in the middle row. Also to ensure
the map generation works, the start has to be on the same row or
muleiplt of 2 squares out. Putting it on the same row just so map
generation gets evenly spread.