Removed arbitrary adjustments of valley walls when placing them
They were originally used to make sure the sprites fitted together
nicely when placed on the map. However it meant that if the sprite was
changed the offsets would need adjusting too. Instead, all sprites are now
anchored in the center of the y-dimension, which means part of each
sprite is above and below where it is placed. This means there is a
standardised position to adhere to with the sprites, and that the
sprites should be changed to fit this pattern.