Allowed enemies to span other enemies upon their destruction
This is a nice feature for enemies to have, as it allows more powerful
enemies to break down into simpler, more familiar ones, rather than
continuing to just get bigger and tougher enemies.

Had to shift around a few functions to make this work, but not too much
change. Now just have to add some enemy types in an array in the enemies
config file.

When the enemy is destroyed, any enemies created are created in a path
backwards from the point of destruction. If there is not enough, space,
add them further down the path. Arbitrarily put 30 steps gap, but will
need to probably change this in future so that it is dependent on the
type of enemies that are spawned.