Implemented bullet piercing
Piercing is now a property of a bullet type, not a tower. Each bullet
type has a number of times it can hit enemies. When down to 0 the bullet
can be removed. Non-piercing is just a limit of 1. In future, if an
upgrade to a piercing, or more enemies for a bullet is chosen, then it
will be the bullet type that is updated.