Introcude react main menu and use webpack for building application
React requires building with babel, so have introduced use of webpack
and babel to do this. This is as an alternative to create-react-app. For
development, run webpack which will compile to the build directoy.

Have removed some of the src in includes from index.html, and switching
to including them in the new index.js.

Added an alternative main menu using react components. Will over time
migrate to using react fully for the menus.