Pass player ID and config to game and start rendering
This a much cleaner way to start the game than previously. Before,
players were addded asynchronously at any time. Now, the players who
will take part in the game are given to the game. If a player leaves,
their score will be greyed out but not removed. This also means the
player ID is no longer injected into the config json which is good.