Render the map as the background to main menu and lobby
This required changing the sprite renderer to not automaticall bind to
the document body. Instead, we use react to add the canvas to add it to
a div which can then be integrated with the rest of the React app.

It is currently fixed as the background, with any map that overfills the
window not getting rendered.