Re-added player info
Each player has a section that has their display name and points
(currently not implemented). It is also in their colour.

To fit all four in, will need to move the lives and money to the right
of the towers, in between tower menu and start game. THen move the base
camp down a bit.