Restructured game view to accomodate multiple players
Space was needd so that the player name/point boxes did not cover the
camp. The camp was originally in the center vertically, but has been
moved down a few rows, and the money and lives counters have been moved
into some free space in the bottom toolbar, which I actually think looks
neater. Their colours have been set to more neutral tones consistent
with the tower menu. This makes the players colours stand out more.