Added initial rounds setup
This lists which enemies will be present each round. The type of enemy are
listed in order as well as the number that will be created, and the
frequency. This will later be parsed and an "enemy HP per second" can be
generated based on the flow of enemies. This can be used to check that
rounds are increasing in difficulty appropriately.