Added script to evaluate difficulty of rounds for balance
This determines how difficult a round will roughly be in terms of the
emeies that are created. It is modelled as a flow of enemies, and
reports both the damage per secomd coming through, and the damage
persecond over a 10 second period. 10s was chosen arbitrarily, but it
felt like enemis would not spend more than 10s in the range of a tower.
As such any more enemies woulld not massively add to the difficulty,
since if it could clear al those in teh 10 period, it could clear any
number of enemies. Likewise if it could not clear them, it would be a
hard round.

Also included the percentage increases of difficulty for each metric.
This shows if the rounds increase in a consisteny manner.

In general, it's quite rough and certainly could be refined. But I think
along side some simulations and play testing, it should hopefully prove
useful.