Rearranged slime enemy sprite dir
Since the colour is the only difference, can just use the same animation
but different colour sprites and exported to another scale. The animation file can be moved to the
relevant directory. It's a bit ad-hoc and not great if later we want to
recreate one of the sprites, but it's a quick and easy way to create the
sprites.