Added the rest of the slime enemies
This was four more, medium, large, giant and uber. Each one had the same
animation, but a different size and colour.

ALso had to make some changes to the game config and bullet mechanics.
Because the slimes are slow, the rounding down of enemy steps meant the
larger and slower slimes had 0 speed. To resolve this, the subgrid size
was increased. This seems to not have affected performance. It did
however slow down all the bullets, as their speed was a fixed value.
This has been changed so that the speed is independent of the subgrid
size.