Started work on the simulator
This simulator will be used to assess the difficulty of the levels and
the game as a whole. It can run through a game very quickly (no visual
output, yet) and so can simulate an entire game in seconds. This can be
done repeatedly to see on average how hard the levels are. Currently it
just reports the lives at the start of each level. These can then be
used to draw a graph of how the lives change over time. A seed can be
provided so that it simulates the same maps each time.

In the future I will add different methods of buying towers, to make it
a more representative sample.

It would also be nice to be able to see visual output of the
simulations. I think will make a separate server endpoint that can be
connected to that immediately starts the game and runs through the
simulations very quickly (but with enough speed to see what is gong on).
It would also need none of the ui other than the map.