Added ability to view simulations
This has been developed mainly as a way to verify the simulations are working as intended, but also as a way to watch specific levels of interest.
It works by sending updates periodically to the client, but still
progressing the level at high speed. Only every few frames is sent
otherwise the client gets overloaded.