Added graphing to display the simulation results
To do this introduces jscharting to show off the simulation results. It
is based off the averages graph, which displays a set of points and then
plots the average through them. I think this is a good way to show the
results for each simulation type, since there will be lots of lines.